Clean up the test suite a bit
The tests were starting to get difficult to work with due to the direct
mocking of stripe API calls.

My attempted solution is to hide the stripe API calls behind a "proxy
object" that can be tested independently and mocked more easily.

Unfortunately as far as I can tell the mocking facilities in
Python/Django aren't very good. I would have preferred to simply create
a fake stripe client that was used at runtime that would return fixed
values.

There is now a StripeTestCase that will mock out all calls to the stripe
gateway class and avoid calls going to the real stripe api (which was
happening)

This commit also moves tests into separate files.
